Fix: Add userId to todos. Fetch/change only for some user

+++ TodoRepository/todoRepository.js
@@ -1,8 +1,8 @@
 import Todo from "../models/todo.js";
 
 class TodosRepository {
-  getAll() {
-    return Todo.find();
+  getAll(userId) {
+    return Todo.find({ userId });
   }
   async create(todo) {
     return Todo.create(todo);
@@ -13,13 +13,13 @@ class TodosRepository {
   delete(id) {
     return Todo.findByIdAndDelete(id);
   }
-  async deleteCompleted() {
-    await Todo.deleteMany({ isDone: true });
-    return Todo.find();
+  async deleteCompleted(userId) {
+    await Todo.deleteMany({ userId, isDone: true });
+    return Todo.find({ userId });
   }
-  async toggleStatus(status) {
-    await Todo.updateMany({ isDone: status });
-    return Todo.find();
+  async toggleStatus(status, userId) {
+    await Todo.updateMany({ userId, isDone: status });
+    return Todo.find({ userId });
   }
 }
